requirements.txt
asgiref==3.3.4
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-cors-headers==3.7.0
django-multiselectfield==0.1.12
django-rest-swagger==2.2.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
PyJWT==2.1.0
pyparsing==3.0.9
pytz==2021.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.1
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==5.2.0
